Skills Auto Complete 1.1
Before adding anew skills whilst there were existing skills did not work. Now it does.

This is because the field skillName on existing skills was a string field and newly added skills had a skillName which was an object. To fix if added a peice of code that recognises if skilName is an object and changes it to a string value