Skills Auto Complete 1.3
Changed Select Component to Creatable Component.

When in Edit State on the profile page and on the Skills tab you can find skills that are already on the system and select them to add to your profile. However the Select Component does not allow you to create new Skills, whereas the Creatable Component does!

Select and Creatable and both part of the react-select https://github.com/JedWatson/react-select